Renaming eventMeta component into eventNewsMeta
