+++ src/shared/components/event-news-meta/index.spec.js
